Make synchronous Rock it's own file
